fix: boto3mod get_connection()
The previous code looks to be a minor change of the boto
`get_connection()` code. In particular, `boto3.session.Session()`
generally does not perform any form of authentication or region
validation, and it certainly does not raise
`boto.exception.NoAuthHandlerFound`.

Have rewritten the code to not attempt to use the old boto 2 module.